Moved org stuff up one level in connector-thegraph
