+++ packages/connector-thegraph/src/parsers/apps.ts
@@ -1,7 +1,7 @@
 import { App as AppDataGql } from '../queries/types'
 import { Organization as OrganizationDataGql } from "../queries/types";
 import { AppData } from 'plumbery-core'
-import { QueryResult } from '../../types'
+import { QueryResult } from '../types'
 
 function _parseApp(app: AppDataGql): AppData {
   return {
