+++ packages/connector-thegraph/src/parsers/repos.ts
@@ -1,7 +1,7 @@
 import { App as AppDataGql } from "../queries/types";
 import { Repo as RepoDataGql } from "../queries/types";
 import { RepoData } from "plumbery-core";
-import { QueryResult } from "../../types";
+import { QueryResult } from "../types";
 
 export function parseRepo(
   result: QueryResult
