+++ packages/connector-thegraph/src/parsers/roles.ts
@@ -1,6 +1,6 @@
 import { Role as RoleDataGql } from "../queries/types";
 import { RoleData } from "plumbery-core";
-import { QueryResult } from "../../types";
+import { QueryResult } from "../types";
 
 export function parseRole(
   result: QueryResult
