+++ packages/connector-thegraph/src/queries/fragments.ts
