+++ packages/connector-thegraph/src/queries/types.ts
